+ Checkpoint    + A functioning MBE up to the third order. But we have to solve the efficiency problem.
 On branch manybody_expansion
 Your branch is up-to-date with 'origin/manybody_expansion'.

 Changes to be committed:
	modified:   ../../interface_to_third_party/molpro_agent.cpp
	modified:   ../compute_expansion_term_periodic_traits.hpp
	modified:   ../compute_interaction_energy_traits.hpp
	modified:   ../report.hpp
	modified:   main.cpp
	modified:   ../../utility/convert_string_to.hpp

 Untracked files:
	../../../bin/
	../../chemistry/test/main.o
	../../chemistry/test/test_element
	../../electron_correlation/client.o
	../../electron_correlation/libiquads_electron_correlation.a
	../../electron_correlation/report.o
	../../electron_correlation/setting.o
	../../electron_correlation/test/client.o
	../../electron_correlation/test/input/
	../../electron_correlation/test/main.o
	../../electron_correlation/test/output/
	../../electron_correlation/test/report.o
	../../electron_correlation/test/scratch/
	../../electron_correlation/test/setting.o
	../../electron_correlation/test/test_electron_correlation
	../../file_system/test/main.o
	../../file_system/test/test_file_system
	../../interface_to_third_party/libiquads_third_party.a
	../../interface_to_third_party/molpro_agent.o
	../../interface_to_third_party/test/molpro_agent.o
	../../interface_to_third_party/test/test_molpro_agent
	../agent.o
	../client.o
	../config.o
	../iquads_manybody_expansion
	../libiquads_manybody_expansion.a
	../main.o
	../report.o
	../request.o
	.exrc
	main.o
	test
	test.11x11x11
	test.13x13x13
	test.3x3x3
	test.5x5x5
	test.7x7x7
	test.9x9x9
	try/
	../../mbe_checkout.sh
	../../structure/test/main.o
	../../structure/test/test
	../../utility/test/main.o
	../../utility/test/test_utility
	../../../test/nonadiabatic_coupling/
	../../../test/numerical_nonadiabatic_coupling/

+++ src/utility/convert_string_to.hpp
@@ -61,10 +61,10 @@ template <> inline float convert_string_to< float > ( std :: string string_value
   { return std :: stof( string_value ); }
 
 template <> inline double convert_string_to< double > ( std :: string string_value )
-  { return std :: stof( string_value ); }
+  { return std :: stod( string_value ); }
 
 template <> inline long double convert_string_to< long double > ( std :: string string_value )
-  { return std :: stof( string_value ); }
+  { return std :: stold( string_value ); }
 
 template <> inline std :: string convert_string_to < std :: string > ( std :: string string_value )
   { return string_value; }
